Implemented University Autocomplete drop down menu
