+++ client/src/components/styles.css
@@ -1,6 +1,76 @@
-.notFound{
-    height: 100%;
+body {
+    font-family: 'Roboto', sans-serif;
+    font-size: 1.1em;
+    font-weight: 400;
+}
+
+.container {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    flex: 1;
+    align-items: center;
+}
+
+.cta {
+    background: linear-gradient(to right, rgba(0, 0, 0, 0.8), rgba(0, 0, 0, 0.1)),
+        url("https://iftin-academy.s3.amazonaws.com/college_student2.jpg") no-repeat right center;
+    /* background: linear-gradient(rgba(0, 0, 0, 0.6), rgba(0, 0, 0, 0.6)), url("https://iftin-academy.s3.amazonaws.com/college-student.jpg") no-repeat right center; */
+    background-size: cover;
+    height: 60vh;
+    width: 100%;
     display: flex;
     align-items: center;
     justify-content: center;
+    margin-bottom: 80px;
+}
+
+.cta h1 {
+    color: #fff;
+    margin-bottom: 20px;
+    font-weight: 300;
+    font-size: 2.5em;
+}
+
+.cta p {
+    color: #fff;
+    font-size: 1.2em;
+    font-weight: 100;
+}
+
+.cta-button {
+    padding: 15px 30px;
+    margin-top: 1em;
+    border-radius: 0%;
+    text-transform: uppercase;
+    font-size: 14px;
+    font-weight: 500;
+    background-color: #fdc500;
+    border-color: #fdc500;
+    color: #000;
+}
+
+.cta-button:hover {
+    background-color: #ffd500;
+    color: #000;
+    border-color: #ffd500;
+}
+
+.cta-button:active {
+    background-color: #ffd500 !important;
+    color: #000 !important;
+    border-color: #ffd500 !important;
+}
+
+.cta-button:focus {
+    background-color: #ffd500 !important;
+    color: #000 !important;
+    border-color: #ffd500 !important;
+    box-shadow: none;
+}
+
+.edu-logo img {
+    width: 100px;
+    margin-top: 17em;
+    margin-right: 5em;
 }
